+++ R/modeling.R
@@ -48,6 +48,14 @@
 #'   regression coefficients (for predicting the mean of the response).
 #'   Default is 5, which makes the prior weakly informative on the
 #'   logit scale.
+#' @param intercept_prior_mean The mean of the Normal distribution prior
+#' for the intercept. By default is NULL, which means the intercept
+#' receives the same prior as `coef_prior_mean`. To zero out the
+#' intercept, set this parameter to 0 and `coef_prior_sd` to a
+#' very small number (0.01 or smaller).
+#' @param intercept_prior_sd The SD of the Normal distribution prior
+#' for the intercept. By default is NULL, which means the intercept
+#' receives the same prior SD as `coef_prior_sd`.
 #' @param phi_prior The mean parameter of the exponential prior on
 #'  phi, which determines the dispersion of the beta distribution. The
 #'  default is .1, which equals a mean of 10 and is thus weakly
@@ -73,8 +81,23 @@
 #'   Only useful if a linear model is being fit to phi.
 #'   Default is 5, which makes the prior weakly informative on the
 #'   exponential scale.
+#' @param phi_intercept_prior_mean The mean of the Normal distribution prior
+#' for the phi (dispersion) regression intercept. By default is NULL,
+#' which means the intercept
+#' receives the same prior as `phi_coef_prior_mean`. To zero out the
+#' intercept, set this parameter to 0 and `phi_coef_prior_sd` to a
+#' very small number (0.01 or smaller).
+#' @param phi_intercept_prior_sd The SD of the Normal distribution prior
+#' for the phi (dispersion) regression intercept. By default is NULL,
+#' which means the intercept
+#' receives the same prior SD as `phi_coef_prior_sd`.
 #' @param extra_prior An additional prior, such as a prior for a specific
-#'  regression coefficient, added to the model by passing one of the `brms`
+#'  regression coefficient, added to the outcome regression by passing one of the `brms`
+#'  functions [brms::set_prior] or [brms::prior_string] with appropriate
+#'  values.
+#'  @param extra_prior_phireg An additional prior, such as a prior for a specific
+#'  regression coefficient, added to the `phi` regression model by
+#'  passing one of the `brms`
 #'  functions [brms::set_prior] or [brms::prior_string] with appropriate
 #'  values.
 #' @param init This parameter is used to determine starting values for
@@ -88,6 +111,8 @@
 #'   more robust starting values (just be sure to scale the covariates so they are
 #'   not too large in absolute size).
 #' @param inits Alias of `init`. Overrides `init` when not NULL.
+#' @param offset Specify a hard-coded offset for the model. Default is 0
+#' (no offset). Also consider putting an informative prior on the intercept instead.
 #' @param ... All other arguments passed on to the `brm` function
 #' @return A `brms` object fitted with the ordered beta regression distribution.
 #' @examples
@@ -130,17 +155,29 @@ ordbetareg <- function(formula=NULL,
                        use_brm_multiple=FALSE,
                        coef_prior_mean=0,
                        coef_prior_sd=5,
+                       intercept_prior_mean=NULL,
+                       intercept_prior_SD=NULL,
                        phi_prior=.1,
                        dirichlet_prior=c(1,1,1),
                        phi_coef_prior_mean=0,
                        phi_coef_prior_sd=5,
+                       phi_intercept_prior_mean=NULL,
+                       phi_intercept_prior_SD=NULL,
                        extra_prior=NULL,
+                       extra_prior_phireg=NULL,
                        init ="0",
                        inits = NULL,
+                       offset=0,
                        ...) {
 
   if(!is.null(inits)){
     init <- inits
+  }
+
+  if(!is.numeric(offset)) {
+
+    stop("If specifying an offset, it should be a number.")
+
   }
 
 
@@ -155,7 +192,7 @@ ordbetareg <- function(formula=NULL,
 
     dv <- all.vars(formula$formula)[1]
 
-    formula$formula <- .update2.formula(formula$formula, "0 + Intercept ")
+    formula$formula <- .update2.formula(formula$formula, paste0(offset," + Intercept "))
 
   } else if('mvbrmsformula' %in% class(formula)) {
 
@@ -169,7 +206,7 @@ ordbetareg <- function(formula=NULL,
 
       if(is.null(var$family)) {
 
-        var$formula <- .update2.formula(var$formula, "0 + Intercept ")
+        var$formula <- .update2.formula(var$formula, paste0(offset," + Intercept "))
 
       }
 
@@ -181,7 +218,7 @@ ordbetareg <- function(formula=NULL,
 
     dv <- all.vars(formula)[1]
 
-    formula <- .update2.formula(formula, "0 + Intercept ")
+    formula <- .update2.formula(formula, paste0(offset," + Intercept "))
 
   }
 
@@ -387,17 +424,26 @@ ordbetareg <- function(formula=NULL,
   #   }
   # } else {
   #
-   }
+  }
+
+  # determine if intercept prior should be specified
+
+
 
 
 
   ordbeta_mod <- .load_ordbetareg(beta_prior=c(coef_prior_mean,
                                                coef_prior_sd),
+                                  intercept_prior=c(intercept_prior_mean,
+                                                    intercept_prior_SD),
                                   phireg_beta_prior = c(phi_coef_prior_mean,
                                                         phi_coef_prior_sd),
+                                  phireg_intercept_prior=c(phi_intercept_prior_mean,
+                                                           phi_intercept_prior_SD),
                                   dirichlet_prior_num=dirichlet_prior,
                                   phi_prior = phi_prior,
                                   extra_prior=extra_prior,
+                                  extra_prior_phireg=extra_prior_phireg,
                                   suffix=suffix)
 
   if('mvbrmsformula' %in% class(formula)) {
@@ -439,39 +485,34 @@ ordbetareg <- function(formula=NULL,
   }
 
 
-
-
-
-
-
   if(use_brm_multiple) {
 
     if(phi_reg) {
 
-      brm_multiple(formula=formula, data=data,
+      out_obj <- brm_multiple(formula=formula, data=data,
                    stanvars=ordbeta_mod$stanvars,
                    family=ordbeta_mod$family,
                    prior=ordbeta_mod$priors_phireg,
-                   init=init,
+                   inits=init,
                    ...)
 
     } else {
 
       if(sep_fam) {
 
-        brm_multiple(formula=formula, data=data,
+        out_obj <- brm_multiple(formula=formula, data=data,
                      stanvars=ordbeta_mod$stanvars,
                      prior=ordbeta_mod$priors,
-                     init=init,
+                     inits=init,
                      ...)
 
       } else {
 
-        brm_multiple(formula=formula, data=data,
+        out_obj <- brm_multiple(formula=formula, data=data,
                      stanvars=ordbeta_mod$stanvars,
                      family=ordbeta_mod$family,
                      prior=ordbeta_mod$priors,
-                     init=init,
+                     inits=init,
                      ...)
 
 
@@ -487,31 +528,31 @@ ordbetareg <- function(formula=NULL,
 
     if(phi_reg) {
 
-      brm(formula=formula, data=data,
+      out_obj <- brm(formula=formula, data=data,
           stanvars=ordbeta_mod$stanvars,
           family=ordbeta_mod$family,
           prior=ordbeta_mod$priors_phireg,
-          init=init,
+          inits=init,
           ...)
 
     } else {
 
       if(sep_fam) {
 
-        brm(formula=formula, data=data,
+        out_obj <- brm(formula=formula, data=data,
             stanvars=ordbeta_mod$stanvars,
             prior=ordbeta_mod$priors,
-            init=init,
+            inits=init,
             ...)
 
 
       } else {
 
-        brm(formula=formula, data=data,
+        out_obj <- brm(formula=formula, data=data,
             stanvars=ordbeta_mod$stanvars,
             family=ordbeta_mod$family,
             prior=ordbeta_mod$priors,
-            init=init,
+            inits=init,
             ...)
 
 
@@ -525,7 +566,9 @@ ordbetareg <- function(formula=NULL,
 
 
 
+    class(out_obj) <- c(class(out_obj),"ordbetareg")
 
+    return(out_obj)
 
 }
 
@@ -533,16 +576,17 @@ ordbetareg <- function(formula=NULL,
 #' Internal Function to Add Ordered Beta Regression Family
 #'
 #' Not exported.
-#' @importFrom brms custom_family stanvar set_prior
+#' @importFrom brms custom_family stanvar set_prior posterior_predict
 #' @noRd
 .load_ordbetareg <- function(beta_prior=NULL,
+                             intercept_prior=NULL,
                              phireg_beta_prior=NULL,
+                             phireg_intercept_prior=NULL,
                              dirichlet_prior_num=NULL,
                              phi_prior=NULL,
                              extra_prior=NULL,
+                             extra_prior_phireg=NULL,
                              suffix="") {
-
-  # function called primarily for its side effects
 
   # custom family
 
@@ -577,6 +621,19 @@ ordbetareg <- function(formula=NULL,
   # For pulling posterior predictions
 
   posterior_predict_ord_beta_reg <- function(i, draws, ...) {
+
+    get_args <- list(...)
+
+    if(!is.null(get_args$ntrys) && get_args$ntrys>5) {
+
+      type <- "continuous"
+
+    } else {
+
+      type <- "combined"
+
+    }
+
     mu <- draws$dpars$mu[, i]
     phi <- draws$dpars$phi
     cutzero <- draws$dpars$cutzero
@@ -596,15 +653,23 @@ ordbetareg <- function(formula=NULL,
       sample(1:3,size=1,prob=c(pr_y0[i],pr_cont[i],pr_y1[i]))
     })
 
-    final_out <- sapply(1:length(outcomes),function(i) {
-      if(outcomes[i]==1) {
-        return(0)
-      } else if(outcomes[i]==2) {
-        return(out_beta[i])
-      } else {
-        return(1)
-      }
-    })
+    if(type=="combined") {
+
+        final_out <- sapply(1:length(outcomes),function(i) {
+          if(outcomes[i]==1) {
+            return(0)
+          } else if(outcomes[i]==2) {
+            return(out_beta[i])
+          } else {
+            return(1)
+          }
+        })
+
+      } else if (type=="continuous") {
+
+        final_out <- out_beta
+
+    }
 
     final_out
 
@@ -767,14 +832,6 @@ ordbetareg <- function(formula=NULL,
       set_prior(paste0("normal(",beta_prior[1],",",beta_prior[2],")"),class="b",resp=substring(suffix[2],2)) +
       set_prior(paste0("exponential(",phi_prior,")"),class="phi",resp=substring(suffix[2],2))
 
-    priors_phireg <- set_prior("normal(0,5)",class="b") +
-      set_prior(paste0("induced_dirichlet([",paste0(dirichlet_prior_num,
-                                                    collapse=","),"]', 0, 1, cutzero, cutone)"),
-                class="cutzero") +
-      set_prior(paste0("induced_dirichlet([",paste0(dirichlet_prior_num,
-                                                    collapse=","),"]', 0, 2, cutzero, cutone)"),
-                class="cutone")
-
   } else {
 
 
@@ -793,7 +850,7 @@ ordbetareg <- function(formula=NULL,
 
   }
 
-  priors_phireg <- set_prior("normal(0,5)",class="b") +
+  priors_phireg <- set_prior("normal(0,5)",class="b",dpar="phi") +
     set_prior(paste0("induced_dirichlet([",paste0(dirichlet_prior_num,
                                                   collapse=","),"]', 0, 1, cutzero, cutone)"),
               class="cutzero") +
@@ -806,9 +863,24 @@ ordbetareg <- function(formula=NULL,
   if(!is.null(extra_prior)) {
 
     priors <- priors + extra_prior
-    priors_phireg <- priors_phireg + extra_prior
+    priors_phireg <- priors_phireg + extra_prior_phireg
 
   }
+
+  if(!is.null(intercept_prior)) {
+
+    priors <- priors + set_prior(paste0("normal(",intercept_prior[1],",",intercept_prior[2],")"),
+                                 coef="Intercept",class="b")
+
+  }
+
+  if(!is.null(phireg_intercept_prior)) {
+
+    priors_phireg <- priors_phireg + set_prior(paste0("normal(",phireg_intercept_prior[1],",",phireg_intercept_prior[2],")"),
+                                               class="Intercept",dpar="phi")
+
+  }
+
 
   # priors <- set_prior("normal(0,5)",class="b") +
   #   prior(constant(0),class="b",coef="Intercept") +